Fix a race condition between the scheduler and physical actions
Before this change, it could happen that a physical action already determines
its next tag, but then the scheduler already advances beyond this tag before the
new event is inserted in the event queue. In other words, determining the tag
and inserting the event at this tag was not an atomic operation.

This change includes the determining of the tag in `schedule_async` while the
lock is held and hence makes scheduling physical actions an atomic operation.