Simple mechanism for collecting statistics during execution
This adds a simple class that can be used to (optionally) collect statistics
about a program execution. This includes information bout the program (e.g.
number of reactors, reactions, posts, etc) and about the execution (e.g. number
of processed events and reactions as well as the number of actions triggered).